move commitStatus to own file
