+++ templates/repo/commit_status.tmpl
@@ -1,4 +1,4 @@
-<!-- make sure this matches the color logic in web_src/js/types.ts -->
+<!-- make sure this matches the color logic in web_src/js/data.ts -->
 {{if eq .State "pending"}}
 	{{svg "octicon-dot-fill" 18 "commit-status icon text yellow"}}
 {{end}}
